Fix css and image edit

+++ app/Http/Controllers/HomestayController.php
@@ -370,6 +370,10 @@ class HomestayController extends Controller
         $roomNo = $request->roomNo;
         $bookings = Booking::where('roomid', $roomId)
                            ->whereIn('status', ['Booked', 'Completed' ,'Deposited' ,'Balance Paid'])
+                           ->where(function($query){
+                                $query->where('checkin','>=' ,date('Y-m-d'))
+                                ->orWhere('checkout','>' ,date('Y-m-d'));
+                           })
                            ->get();
         
         $disabledDates = [];
@@ -1078,16 +1082,18 @@ public function getPromotionHistory(Request $request){
                 ->pluck('image_path')
                 ->first();
                 // delete the image
-                unlink(public_path($deletePath));
+                if(public_path($deletePath)){
+                    unlink(public_path($deletePath));
+                }
                 DB::table('homestay_images')
                 ->where('id',$id)
                 ->delete();
 
             }            
         }
-
         // for updating images
         $editIds = $request->edit_id;
+
         if($request->file('image') != null && $editIds != null){
             $editIdsArray = explode(',', $editIds);//like split
             sort($editIdsArray);
@@ -1102,7 +1108,9 @@ public function getPromotionHistory(Request $request){
                     ->where('id', $id)
                     ->first();
                     // delete the old image at this position
-                    unlink(public_path($image->image_path));
+                    if(public_path($image->image_path)){
+                        unlink(public_path($image->image_path));
+                    }
                     // place the new one
                     $pathInfo = pathinfo($image->image_path);
                     $newImagePath = 'homestay-image/'.$pathInfo['filename'].".".$newImages[$key]->getClientOriginalExtension();
